+++ src/Command/DeleteStorageHandler.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Command;
 
+use App\Exception\StorageCannotBeDeleted;
 use App\Repository\StorageRepository;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
 
@@ -18,6 +19,15 @@ final readonly class DeleteStorageHandler
     public function __invoke(DeleteStorageCommand $command): void
     {
         $storage = $this->storageRepository->get($command->storageId);
+
+        if ($storage->isOccupied()) {
+            throw StorageCannotBeDeleted::becauseItIsOccupied($storage);
+        }
+
+        if ($storage->isReserved()) {
+            throw StorageCannotBeDeleted::becauseItIsReserved($storage);
+        }
+
         $this->storageRepository->delete($storage);
     }
 }
